bot/requirements.txt
boto3==1.34.126
aiogram==3.7.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
Pillow==10.3.0
ultralytics
numpy==1.26.4
ensemble-boxes==1.0.9
python-dotenv
fastapi==0.111.0